+++ packages/database-context/src/IDatabaseContext.ts
@@ -1,3 +1,0 @@
-export default interface IDatabaseContext {
-    runWithTransaction<T>(fn: () => Promise<T>): Promise<T>;
-}