+++ packages/population/src/index.ts
@@ -1,3 +1,0 @@
-export * from "./NormalizedPopulate.js";
-export * from "./normalize.js";
-export * from "./Populate.js";