+++ packages/prisma-expand/src/index.ts
@@ -1,0 +1,2 @@
+export { default as BasePrismaExpander } from "./BasePrismaExpander.js";
+export { type PrismaInclude, toPrismaInclude } from "./toPrismaInclude.js";