+++ packages/unit-of-work/src/IUnitOfWork.ts
@@ -1,0 +1,3 @@
+export default interface IUnitOfWork {
+    runWithTransaction<T>(fn: () => Promise<T>): Promise<T>;
+}